switch to book insead of sparkles
